+++ CHANGELOG.md
@@ -1,1 +1,5 @@
-Initial Release
+Yay some simple features
+
+## New features
+- Added a resize() method which allows you to resize the editor
+- Added a on_input() method which allows you to hook into the input event of the editor